vision/target_finder/requirements.txt
requests
Pillow
torch
transformers
safetensors
scipy

